Add window confirmation for deleting payments.

+++ app/scripts/controllers/paymentModal.js
@@ -176,9 +176,11 @@ angular.module('confRegistrationWebApp')
     };
 
     $scope.deletePayment = function (payment) {
-      $http.delete('payments/' + payment.id, payment).success(function() {
-        loadPayments();
-      });
+      if(window.confirm("Do you really want to delete this payment?")) {
+        $http.delete('payments/' + payment.id, payment).success(function () {
+          loadPayments();
+        });
+      }
     };
 
   });
